backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
alembic>=1.13.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
celery[redis]>=5.3.0
redis>=5.0.0
flower>=2.0.0
cerebras-cloud-sdk>=1.0.0
python-dotenv>=1.0.0
